+++ packages/lwc-engine/jest.config.js
@@ -3,7 +3,10 @@ const BASE_CONFIG = require('../../scripts/jest/base.config');
 
 module.exports = {
     ...BASE_CONFIG,
-    displayName: 'lwc-engine',
 
+    displayName: 'lwc-engine',
+    moduleNameMapper: {
+        'test-utils': path.resolve(__dirname, 'scripts/jest/test-utils.js'),
+    },
     setupTestFrameworkScriptFile: path.resolve(__dirname, 'scripts/jest/setup-test.js'),
 };