+++ packages/lwc-engine/src/framework/__tests__/html-element.spec.ts
@@ -1,3 +1,5 @@
+import { compileTemplate } from 'test-utils';
+
 import { createElement, register, unwrap } from '../main';
 import { getHostShadowRoot, LightningElement } from '../html-element';
 import assertLogger from '../../shared/assert';
@@ -12,11 +14,19 @@ describe('html-element', () => {
             }
             Child.publicMethods = ['setFoo'];
 
+            const html = compileTemplate(`
+                <template>
+                    <x-child></x-child>
+                </template>
+            `, {
+                modules: {
+                    'x-child': Child
+                }
+            });
+
             class Parent extends LightningElement {
                 render() {
-                    return ($api) => {
-                        return [$api.c('x-child', Child, {})]
-                    }
+                    return html;
                 }
             }
             const element = createElement('should-set-attribute-on-host-element-when-element-is-nested-in-template', { is: Parent });
@@ -670,15 +680,34 @@ describe('html-element', () => {
         // TODO: This test log multiple errors. We should fix this before migrating to expect().toLogError()
         it('should log error message when attribute is set via elm.setAttribute if reflective property is defined', () => {
             jest.spyOn(assertLogger, 'logError');
-            class MyComponent extends LightningElement {}
-            const elm = createElement('x-foo', {is: MyComponent});
-            elm.setAttribute('tabindex', '0');
+            class Child extends LightningElement {}
+            function html($api, $cmp) {
+                return [
+                    $api.c('x-child', Child, { attrs: { title: 'child title' }})
+                ];
+            }
+            class Parent extends LightningElement {
+                render() {
+                    return html;
+                }
+                renderedCallback() {
+                    this.template.querySelector('x-child').setAttribute('tabindex', 0);
+                }
+            }
+            const elm = createElement('x-foo', {is: Parent});
             document.body.appendChild(elm);
 
             return Promise.resolve().then( () => {
                 expect(assertLogger.logError).toBeCalled();
                 assertLogger.logError.mockRestore();
             });
+        });
+
+        it('should not throw when accessing attribute in root elements', () => {
+            class Parent extends LightningElement {}
+            const elm = createElement('x-foo', {is: Parent});
+            document.body.appendChild(elm);
+            elm.setAttribute('tabindex', 1);
         });
 
         it('should delete existing attribute prior rendering', () => {