+++ packages/lwc-engine/src/framework/modules/attrs.ts
@@ -1,5 +1,5 @@
 import assert from "../../shared/assert";
-import { prepareForValidAttributeMutation } from '../restrictions';
+import { unlockAttribute, lockAttribute } from '../restrictions';
 import { isUndefined, keys, StringCharCodeAt, isNull } from '../../shared/language';
 import { EmptyObject } from '../utils';
 import { Module, VNode } from "../../3rdparty/snabbdom/types";
@@ -35,7 +35,7 @@ function updateAttrs(oldVnode: VNode, vnode: VNode) {
         const old = (oldAttrs as any)[key];
         if (old !== cur) {
             if (process.env.NODE_ENV !== 'production') {
-                prepareForValidAttributeMutation(elm, key);
+                unlockAttribute(elm, key);
             }
             if (StringCharCodeAt.call(key, 3) === ColonCharCode) {
                 // Assume xml namespace
@@ -47,6 +47,9 @@ function updateAttrs(oldVnode: VNode, vnode: VNode) {
                 elm.removeAttribute(key);
             } else {
                 elm.setAttribute(key, cur as string);
+            }
+            if (process.env.NODE_ENV !== 'production') {
+                lockAttribute(elm, key);
             }
         }
     }